Added Prim algorithm implementation

+++ SDIZO/include/structures/Edge.hpp
@@ -4,14 +4,14 @@ namespace SDIZO
 {
 	struct Edge
 	{
-		Edge(int origin, int destination, int value)
+		Edge(size_t origin, size_t destination, size_t value)
 			: origin(origin), destination(destination), 
 			value(value), previous(nullptr), next(nullptr) {}
 
 		Edge* previous;
 		Edge* next;
-		int origin;
-		int destination;
-		int value;
+		size_t origin;
+		size_t destination;
+		size_t value;
 	};
 }
